Updated storage.pm to work with larger volumes
In reference to:
https://github.com/centreon/centreon-plugins/issues/1651
https://github.com/centreon/centreon-plugins/issues/983

Since the Windows SNMP agent utilizes signed 32-bit Integer variables for storing the size and used amounts of disks, they are prone to overflow and report negative values when working with large disks. This simple workaround checks for a negative return value and adds the 32-bit range (2^32). From personal experience, this works with disks of up to 40TB. I am however not sure of the actual limit this works with.