fix(java/awa): agent mode - show agents having empty date in LastCheck
Some agents have a `LastCheck` field with only the time part (no date part),
and in this case the only time value I saw for LastCheck was `00:00:00`.

The current code will remove those agents from the list, however I believe
there is a need for monitoring those agents (those never
were connected, but are still defined in the system).

This patch allows those agents to be listed, and (arbitrarily) sets the
lastcheck time difference (field: `since`) to `0` - which may, or may not, be
problematic. (We could set it to a date far in the past, but that would
be strange on a graph. We cannot remove the field as it is necessary
for the display)